dummy_data/agent_service/requirements.txt
duckdb>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.25.0
pydantic>=2.0.0
pyyaml>=6.0.0
requests>=2.31.0
